+++ lib/stream/manager/from_directory.go
@@ -14,6 +14,8 @@ import (
 
 // LoadStreamConfigsFromDirectory reads a map of stream ids to configurations
 // by walking a directory of .json and .yaml files.
+//
+// Deprecated: Use LoadStreamConfigsFromPath instead.
 func LoadStreamConfigsFromDirectory(replaceEnvVars bool, dir string) (map[string]stream.Config, error) {
 	streamMap := map[string]stream.Config{}